Move `ActionDispatch::Mime` to `ActiveSupport::Mime`
This will now make ActionView independent of ActionDispatch (and thus
ActionPack). This also makes Rails' Mime code more easily usable by
other people!

ref #17030
