+++ actionmailer/lib/action_mailer.rb
@@ -50,9 +50,9 @@ module ActionMailer
   autoload :DeliveryJob
 end
 
-autoload :Mime, 'action_dispatch/http/mime_type'
+autoload :Mime, 'active_support/mime'
 
 ActiveSupport.on_load(:action_view) do
-  ActionView::Base.default_formats ||= Mime::SET.symbols
-  ActionView::Template::Types.delegate_to Mime
+  ActionView::Base.default_formats ||= ActiveSupport::Mime::SET.symbols
+  ActionView::Template::Types.delegate_to ActiveSupport::Mime
 end
