+++ actionpack/lib/abstract_controller/collector.rb
@@ -1,4 +1,4 @@
-require "action_dispatch/http/mime_type"
+require "active_support/mime"
 
 module AbstractController
   module Collector
@@ -6,23 +6,23 @@ module AbstractController
       sym = mime.is_a?(Symbol) ? mime : mime.to_sym
       class_eval <<-RUBY, __FILE__, __LINE__ + 1
         def #{sym}(*args, &block)
-          custom(Mime[:#{sym}], *args, &block)
+          custom(ActiveSupport::Mime[:#{sym}], *args, &block)
         end
       RUBY
     end
 
-    Mime::SET.each do |mime|
+    ActiveSupport::Mime::SET.each do |mime|
       generate_method_for_mime(mime)
     end
 
-    Mime::Type.register_callback do |mime|
+    ActiveSupport::Mime::Type.register_callback do |mime|
       generate_method_for_mime(mime) unless self.instance_methods.include?(mime.to_sym)
     end
 
   protected
 
     def method_missing(symbol, &block)
-      unless mime_constant = Mime[symbol]
+      unless mime_constant = ActiveSupport::Mime[symbol]
         raise NoMethodError, "To respond to a custom format, register it as a MIME type first: " \
           "http://guides.rubyonrails.org/action_controller_overview.html#restful-downloads. " \
           "If you meant to respond to a variant like :tablet or :phone, not a custom format, " \
@@ -30,7 +30,7 @@ module AbstractController
           "format.html { |html| html.tablet { ... } }"
       end
 
-      if Mime::SET.include?(mime_constant)
+      if ActiveSupport::Mime::SET.include?(mime_constant)
         AbstractController::Collector.generate_method_for_mime(mime_constant)
         send(symbol, &block)
       else
