+++ actionpack/lib/action_controller/metal/rendering.rb
@@ -63,7 +63,7 @@ module ActionController
     end
 
     def _set_html_content_type
-      self.content_type = Mime[:html].to_s
+      self.content_type = ActiveSupport::Mime[:html].to_s
     end
 
     def _set_rendered_content_type(format)
