+++ actionpack/test/controller/action_pack_assertions_test.rb
@@ -65,7 +65,7 @@ class ActionPackAssertionsController < ActionController::Base
   end
 
   def render_text_with_custom_content_type
-    render body: "Hello!", content_type: Mime[:rss]
+    render body: "Hello!", content_type: ActiveSupport::Mime[:rss]
   end
 
   def session_stuffing
