+++ actionpack/test/controller/content_type_test.rb
@@ -3,7 +3,7 @@ require 'abstract_unit'
 class OldContentTypeController < ActionController::Base
   # :ported:
   def render_content_type_from_body
-    response.content_type = Mime[:rss]
+    response.content_type = ActiveSupport::Mime[:rss]
     render body: "hello world!"
   end
 
@@ -14,7 +14,7 @@ class OldContentTypeController < ActionController::Base
 
   # :ported:
   def render_content_type_from_render
-    render body: "hello world!", :content_type => Mime[:rss]
+    render body: "hello world!", :content_type => ActiveSupport::Mime[:rss]
   end
 
   # :ported:
@@ -36,7 +36,7 @@ class OldContentTypeController < ActionController::Base
   end
 
   def render_change_for_builder
-    response.content_type = Mime[:html]
+    response.content_type = ActiveSupport::Mime[:html]
     render :action => "render_default_for_builder"
   end
 
@@ -45,7 +45,7 @@ class OldContentTypeController < ActionController::Base
       format.html { render body: "hello world!" }
       format.xml  { render action: "render_default_content_types_for_respond_to" }
       format.js   { render body: "hello world!" }
-      format.rss  { render body: "hello world!", content_type: Mime[:xml] }
+      format.rss  { render body: "hello world!", content_type: ActiveSupport::Mime[:xml] }
     end
   end
 end
@@ -64,68 +64,68 @@ class ContentTypeTest < ActionController::TestCase
   def test_render_defaults
     get :render_defaults
     assert_equal "utf-8", @response.charset
-    assert_equal Mime[:text], @response.content_type
+    assert_equal ActiveSupport::Mime[:text], @response.content_type
   end
 
   def test_render_changed_charset_default
     with_default_charset "utf-16" do
       get :render_defaults
       assert_equal "utf-16", @response.charset
-      assert_equal Mime[:text], @response.content_type
+      assert_equal ActiveSupport::Mime[:text], @response.content_type
     end
   end
 
   # :ported:
   def test_content_type_from_body
     get :render_content_type_from_body
-    assert_equal Mime[:rss], @response.content_type
+    assert_equal ActiveSupport::Mime[:rss], @response.content_type
     assert_equal "utf-8", @response.charset
   end
 
   # :ported:
   def test_content_type_from_render
     get :render_content_type_from_render
-    assert_equal Mime[:rss], @response.content_type
+    assert_equal ActiveSupport::Mime[:rss], @response.content_type
     assert_equal "utf-8", @response.charset
   end
 
   # :ported:
   def test_charset_from_body
     get :render_charset_from_body
-    assert_equal Mime[:text], @response.content_type
+    assert_equal ActiveSupport::Mime[:text], @response.content_type
     assert_equal "utf-16", @response.charset
   end
 
   # :ported:
   def test_nil_charset_from_body
     get :render_nil_charset_from_body
-    assert_equal Mime[:text], @response.content_type
+    assert_equal ActiveSupport::Mime[:text], @response.content_type
     assert_equal "utf-8", @response.charset, @response.headers.inspect
   end
 
   def test_nil_default_for_erb
     with_default_charset nil do
       get :render_default_for_erb
-      assert_equal Mime[:html], @response.content_type
+      assert_equal ActiveSupport::Mime[:html], @response.content_type
       assert_nil @response.charset, @response.headers.inspect
     end
   end
 
   def test_default_for_erb
     get :render_default_for_erb
-    assert_equal Mime[:html], @response.content_type
+    assert_equal ActiveSupport::Mime[:html], @response.content_type
     assert_equal "utf-8", @response.charset
   end
 
   def test_default_for_builder
     get :render_default_for_builder
-    assert_equal Mime[:xml], @response.content_type
+    assert_equal ActiveSupport::Mime[:xml], @response.content_type
     assert_equal "utf-8", @response.charset
   end
 
   def test_change_for_builder
     get :render_change_for_builder
-    assert_equal Mime[:html], @response.content_type
+    assert_equal ActiveSupport::Mime[:html], @response.content_type
     assert_equal "utf-8", @response.charset
   end
 
@@ -144,24 +144,24 @@ class AcceptBasedContentTypeTest < ActionController::TestCase
   tests OldContentTypeController
 
   def test_render_default_content_types_for_respond_to
-    @request.accept = Mime[:html].to_s
+    @request.accept = ActiveSupport::Mime[:html].to_s
     get :render_default_content_types_for_respond_to
-    assert_equal Mime[:html], @response.content_type
+    assert_equal ActiveSupport::Mime[:html], @response.content_type
 
-    @request.accept = Mime[:js].to_s
+    @request.accept = ActiveSupport::Mime[:js].to_s
     get :render_default_content_types_for_respond_to
-    assert_equal Mime[:js], @response.content_type
+    assert_equal ActiveSupport::Mime[:js], @response.content_type
   end
 
   def test_render_default_content_types_for_respond_to_with_template
-    @request.accept = Mime[:xml].to_s
+    @request.accept = ActiveSupport::Mime[:xml].to_s
     get :render_default_content_types_for_respond_to
-    assert_equal Mime[:xml], @response.content_type
+    assert_equal ActiveSupport::Mime[:xml], @response.content_type
   end
 
   def test_render_default_content_types_for_respond_to_with_overwrite
-    @request.accept = Mime[:rss].to_s
+    @request.accept = ActiveSupport::Mime[:rss].to_s
     get :render_default_content_types_for_respond_to
-    assert_equal Mime[:xml], @response.content_type
+    assert_equal ActiveSupport::Mime[:xml], @response.content_type
   end
 end
