+++ actionpack/test/controller/new_base/content_type_test.rb
@@ -7,12 +7,12 @@ module ContentType
     end
 
     def set_on_response_obj
-      response.content_type = Mime[:rss]
+      response.content_type = ActiveSupport::Mime[:rss]
       render body: "Hello world!"
     end
 
     def set_on_render
-      render body: "Hello world!", content_type: Mime[:rss]
+      render body: "Hello world!", content_type: ActiveSupport::Mime[:rss]
     end
   end
