+++ actionpack/test/dispatch/request_test.rb
@@ -754,33 +754,33 @@ class RequestFormat < BaseRequestTest
   test "xml format" do
     request = stub_request
     assert_called(request, :parameters, times: 2, returns: {format: :xml}) do
-      assert_equal Mime[:xml], request.format
+      assert_equal ActiveSupport::Mime[:xml], request.format
     end
   end
 
   test "xhtml format" do
     request = stub_request
     assert_called(request, :parameters, times: 2, returns: {format: :xhtml}) do
-      assert_equal Mime[:html], request.format
+      assert_equal ActiveSupport::Mime[:html], request.format
     end
   end
 
   test "txt format" do
     request = stub_request
     assert_called(request, :parameters, times: 2, returns: {format: :txt}) do
-      assert_equal Mime[:text], request.format
+      assert_equal ActiveSupport::Mime[:text], request.format
     end
   end
 
   test "XMLHttpRequest" do
     request = stub_request(
       'HTTP_X_REQUESTED_WITH' => 'XMLHttpRequest',
-      'HTTP_ACCEPT' => [Mime[:js], Mime[:html], Mime[:xml], "text/xml", "*/*"].join(",")
+      'HTTP_ACCEPT' => [ActiveSupport::Mime[:js], ActiveSupport::Mime[:html], ActiveSupport::Mime[:xml], "text/xml", "*/*"].join(",")
     )
 
     assert_called(request, :parameters, times: 1, returns: {}) do
       assert request.xhr?
-      assert_equal Mime[:js], request.format
+      assert_equal ActiveSupport::Mime[:js], request.format
     end
   end
 
@@ -800,36 +800,36 @@ class RequestFormat < BaseRequestTest
 
   test "formats text/html with accept header" do
     request = stub_request 'HTTP_ACCEPT' => 'text/html'
-    assert_equal [Mime[:html]], request.formats
+    assert_equal [ActiveSupport::Mime[:html]], request.formats
   end
 
   test "formats blank with accept header" do
     request = stub_request 'HTTP_ACCEPT' => ''
-    assert_equal [Mime[:html]], request.formats
+    assert_equal [ActiveSupport::Mime[:html]], request.formats
   end
 
   test "formats XMLHttpRequest with accept header" do
     request = stub_request 'HTTP_X_REQUESTED_WITH' => "XMLHttpRequest"
-    assert_equal [Mime[:js]], request.formats
+    assert_equal [ActiveSupport::Mime[:js]], request.formats
   end
 
   test "formats application/xml with accept header" do
     request = stub_request('CONTENT_TYPE' => 'application/xml; charset=UTF-8',
                            'HTTP_X_REQUESTED_WITH' => "XMLHttpRequest")
-    assert_equal [Mime[:xml]], request.formats
+    assert_equal [ActiveSupport::Mime[:xml]], request.formats
   end
 
   test "formats format:text with accept header" do
     request = stub_request
     assert_called(request, :parameters, times: 2, returns: {format: :txt}) do
-      assert_equal [Mime[:text]], request.formats
+      assert_equal [ActiveSupport::Mime[:text]], request.formats
     end
   end
 
   test "formats format:unknown with accept header" do
     request = stub_request
     assert_called(request, :parameters, times: 2, returns: {format: :unknown}) do
-      assert_instance_of Mime::NullType, request.format
+      assert_instance_of ActiveSupport::Mime::NullType, request.format
     end
   end
 
@@ -852,7 +852,7 @@ class RequestFormat < BaseRequestTest
   test "formats with xhr request" do
     request = stub_request 'HTTP_X_REQUESTED_WITH' => "XMLHttpRequest"
     assert_called(request, :parameters, times: 1, returns: {}) do
-      assert_equal [Mime[:js]], request.formats
+      assert_equal [ActiveSupport::Mime[:js]], request.formats
     end
   end
 
@@ -863,35 +863,35 @@ class RequestFormat < BaseRequestTest
     begin
       request = stub_request 'HTTP_ACCEPT' => 'application/xml'
       assert_called(request, :parameters, times: 1, returns: {}) do
-        assert_equal [ Mime[:html] ], request.formats
+        assert_equal [ ActiveSupport::Mime[:html] ], request.formats
       end
 
       request = stub_request 'HTTP_ACCEPT' => 'koz-asked/something-crazy'
       assert_called(request, :parameters, times: 1, returns: {}) do
-        assert_equal [ Mime[:html] ], request.formats
+        assert_equal [ ActiveSupport::Mime[:html] ], request.formats
       end
 
       request = stub_request 'HTTP_ACCEPT' => '*/*;q=0.1'
       assert_called(request, :parameters, times: 1, returns: {}) do
-        assert_equal [ Mime[:html] ], request.formats
+        assert_equal [ ActiveSupport::Mime[:html] ], request.formats
       end
 
       request = stub_request 'HTTP_ACCEPT' => 'application/jxw'
       assert_called(request, :parameters, times: 1, returns: {}) do
-        assert_equal [ Mime[:html] ], request.formats
+        assert_equal [ ActiveSupport::Mime[:html] ], request.formats
       end
 
       request = stub_request 'HTTP_ACCEPT' => 'application/xml',
                              'HTTP_X_REQUESTED_WITH' => "XMLHttpRequest"
 
       assert_called(request, :parameters, times: 1, returns: {}) do
-        assert_equal [ Mime[:js] ], request.formats
+        assert_equal [ ActiveSupport::Mime[:js] ], request.formats
       end
 
       request = stub_request 'HTTP_ACCEPT' => 'application/xml',
                              'HTTP_X_REQUESTED_WITH' => "XMLHttpRequest"
       assert_called(request, :parameters, times: 2, returns: {format: :json}) do
-        assert_equal [ Mime[:json] ], request.formats
+        assert_equal [ ActiveSupport::Mime[:json] ], request.formats
       end
     ensure
       ActionDispatch::Request.ignore_accept_header = old_ignore_accept_header
@@ -901,7 +901,7 @@ end
 
 class RequestMimeType < BaseRequestTest
   test "content type" do
-    assert_equal Mime[:html], stub_request('CONTENT_TYPE' => 'text/html').content_mime_type
+    assert_equal ActiveSupport::Mime[:html], stub_request('CONTENT_TYPE' => 'text/html').content_mime_type
   end
 
   test "no content type" do
@@ -909,11 +909,11 @@ class RequestMimeType < BaseRequestTest
   end
 
   test "content type is XML" do
-    assert_equal Mime[:xml], stub_request('CONTENT_TYPE' => 'application/xml').content_mime_type
+    assert_equal ActiveSupport::Mime[:xml], stub_request('CONTENT_TYPE' => 'application/xml').content_mime_type
   end
 
   test "content type with charset" do
-    assert_equal Mime[:xml], stub_request('CONTENT_TYPE' => 'application/xml; charset=UTF-8').content_mime_type
+    assert_equal ActiveSupport::Mime[:xml], stub_request('CONTENT_TYPE' => 'application/xml; charset=UTF-8').content_mime_type
   end
 
   test "user agent" do
@@ -926,9 +926,9 @@ class RequestMimeType < BaseRequestTest
       'HTTP_X_REQUESTED_WITH' => "XMLHttpRequest"
     )
 
-    assert_equal nil, request.negotiate_mime([Mime[:xml], Mime[:json]])
-    assert_equal Mime[:html], request.negotiate_mime([Mime[:xml], Mime[:html]])
-    assert_equal Mime[:html], request.negotiate_mime([Mime[:xml], Mime::ALL])
+    assert_equal nil, request.negotiate_mime([ActiveSupport::Mime[:xml], ActiveSupport::Mime[:json]])
+    assert_equal ActiveSupport::Mime[:html], request.negotiate_mime([ActiveSupport::Mime[:xml], ActiveSupport::Mime[:html]])
+    assert_equal ActiveSupport::Mime[:html], request.negotiate_mime([ActiveSupport::Mime[:xml], ActiveSupport::Mime::ALL])
   end
 
   test "negotiate_mime with content_type" do
@@ -937,7 +937,7 @@ class RequestMimeType < BaseRequestTest
       'HTTP_X_REQUESTED_WITH' => "XMLHttpRequest"
     )
 
-    assert_equal Mime[:xml], request.negotiate_mime([Mime[:xml], Mime[:csv]])
+    assert_equal ActiveSupport::Mime[:xml], request.negotiate_mime([ActiveSupport::Mime[:xml], ActiveSupport::Mime[:csv]])
   end
 end
